requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
cfgv==3.4.0
colorama==0.4.6; platform_system == "Windows"
distlib==0.3.8
distro==1.9.0
filelock==3.14.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
identify==2.5.36
idna==3.7
nodeenv==1.8.0
openai==1.26.0
platformdirs==4.2.1
pre-commit==3.7.0
pydantic==2.7.1
pydantic-core==2.18.2
python-dotenv==1.0.1
pyyaml==6.0.1
setuptools==69.5.1
sniffio==1.3.1
tqdm==4.66.4
typing-extensions==4.11.0
virtualenv==20.26.1
